requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
crispy-bootstrap5==2024.2
Django==4.2.15
django-bootstrap-v5==1.0.11
django-cors-headers==4.4.0
django-crispy-forms==2.3
djangorestframework==3.15.2
Markdown==3.5.2
numpy==2.0.1
opencv-python==4.10.0.84
pillow==10.4.0
soupsieve==2.6
sqlparse==0.5.0
tzdata==2024.1
